created /dashboards endpoint to manage dashboard page, enabled SSR for Electricity Dashboard
