+++ sites/geohub/src/routes/(app)/maps/+page.ts
@@ -3,7 +3,7 @@ import type { PageLoad } from './$types';
 export const load: PageLoad = async ({ data }) => {
 	const { styles } = data;
 	const title = 'Maps | GeoHub';
-	const content = 'Maps';
+	const content = 'Explore maps';
 
 	return {
 		title,
