+++ sites/geohub/src/routes/(map)/dashboards/electricity/+page.ts
@@ -12,6 +12,3 @@ export const load: PageLoad = async ({ data }) => {
 		azureUrl
 	};
 };
-
-export const csr = true;
-export const ssr = false;
